move to plugin based server configuration
