+++ tooling/cli/src/commands/codegen.ts
@@ -8,7 +8,7 @@ import { defineCommand } from "citty";
 import path from "pathe";
 
 import { logger } from "../common";
-import { isArriConfig, ResolvedArriConfig } from "../config";
+import { ArriConfig, isArriConfig } from "../config";
 
 export default defineCommand({
     meta: {
@@ -147,7 +147,7 @@ async function getAppDefinitionFromFile(file: string) {
 
 async function runGenerators(
     def: AppDefinition,
-    generators: ResolvedArriConfig["generators"],
+    generators: ArriConfig["generators"],
 ) {
     logger.info(`Generating ${generators?.length} client(s)`);
     await Promise.allSettled(
