+++ tooling/cli/src/serverConfigs/_index.ts
@@ -1,0 +1,1 @@
+export * from "./tsServer";
