+++ tooling/cli/src/serverConfigs/tsServer.test.ts
@@ -1,16 +1,15 @@
-import { getRpcMetaFromPath } from "../common";
-import { type ResolvedArriConfig } from "../config";
+import { TsServerConfig } from "../_index";
+import { getRpcMetaFromPath } from "./tsServer";
 
 describe("Naming RPCs", () => {
     test("Basic route", () => {
-        const config: ResolvedArriConfig = {
+        const config: Required<TsServerConfig> = {
             port: 3000,
             rootDir: "/files/items/examples-app",
             srcDir: "src",
             entry: "",
             procedureDir: "procedures",
             procedureGlobPatterns: ["**/*.rpc.ts"],
-            generators: [],
             buildDir: ".arri",
             esbuild: {},
             serverEntry: "",
@@ -26,14 +25,13 @@ describe("Naming RPCs", () => {
         expect(result?.httpPath).toBe("/users/get-user");
     });
     test("Route with weird chars", () => {
-        const config: ResolvedArriConfig = {
+        const config: Required<TsServerConfig> = {
             port: 3000,
             rootDir: "",
             srcDir: "src",
             entry: "",
             procedureDir: "procedures",
             procedureGlobPatterns: ["**/*.rpc.ts"],
-            generators: [],
             buildDir: ".arri",
             esbuild: {},
             serverEntry: "",
